+++ Source Code Inspection/src/main/java/br/calebe/ticketmachine/core/PapelMoeda.java
@@ -1,0 +1,24 @@
+package br.calebe.ticketmachine.core;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class PapelMoeda {
+
+    protected int valor;
+    protected int quantidade;
+
+    public PapelMoeda(int valor, int quantidade) {
+        this.valor = valor;
+        this.quantidade = quantidade;
+    }
+
+    public int getValor() {
+        return valor;
+    }
+
+    public int getQuantidade() {
+        return quantidade;
+    }
+}
