+++ Source Code Inspection/src/main/java/br/calebe/ticketmachine/exception/PapelMoedaInvalidaException.java
@@ -1,0 +1,9 @@
+package br.calebe.ticketmachine.exception;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class PapelMoedaInvalidaException extends Exception {
+    
+}
