+++ Source Code Inspection/src/main/java/br/calebe/ticketmachine/exception/SaldoInsuficienteException.java
@@ -1,0 +1,8 @@
+package br.calebe.ticketmachine.exception;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class SaldoInsuficienteException extends Exception {
+}
